requirements.txt
Flask==2.3.3
mysql-connector-python==9.4.0
Werkzeug==2.3.8